# analysisapi/api-docs/analysisapi/api-specification.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Analysis API",
    "description": "Analysis API の説明です。",
    "version": "1.0.0"
  },
  "paths": {
    "/api/user/{user_id}/{diary_id}": {
      "get": {
        "tags": [
          "analysisapi"
        ],
        "summary": "Get Diary Add Db",
        "description": "指定idのユーザーの指定idの日記を取得し、ベクトルDBに追加します。\n\nArgs:\n  user_id (int): ユーザーID。\n  diary_id (int): 日記ID。\n\nReturns:\n  dict: 日記が正常にDBに追加されたことを示すメッセージ。\n\nRaises:\n  HTTPException: サーバーからのHTTPエラー応答が発生した場合。\n  RequestException: 通信中にエラーが発生した場合。\n  Exception: その他の予期しないエラー。",
        "operationId": "get_diary_add_db_api_user__user_id___diary_id__get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          },
          {
            "name": "diary_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Diary Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Diary Add Db Api User  User Id   Diary Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/analysis/scheduler/{user_id}": {
      "get": {
        "tags": [
          "analysisapi"
        ],
        "summary": "Get Schedule",
        "description": "過去の日記を分析して、明日の予定を提案します。\n\nArgs:\n  user_id (int): ユーザーID。\n\nReturns:\n  str: 明日の予定。",
        "operationId": "get_schedule_api_analysis_scheduler__user_id__get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "title": "Response Get Schedule Api Analysis Scheduler  User Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}